feat: add FpsCanvas component with game logic and dynamic imports; implement prebuild checks for page safety
