+++ .lighthouserc.json
@@ -1,0 +1,10 @@
+{
+  "ci": {
+    "collect": {
+      "url": ["http://localhost:3000"]
+    },
+    "assert": {
+      "preset": "lighthouse:recommended"
+    }
+  }
+}
