+++ app/package.json
@@ -11,7 +11,8 @@
   },
   "scripts": {
     "dev": "next dev",
-    "build": "next build",
+  "build": "next build",
+  "prebuild": "node ./scripts/prebuild-check.js",
     "start": "next start",
   "test": "jest --watch",
   "test:ci": "jest --ci --runInBand --colors",
